backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
httpx>=0.25.0
deepgram-sdk>=3.0.0
groq>=0.4.0
elevenlabs>=1.0.0
